+++ CHANGELOG.md
@@ -9,7 +9,33 @@ Breaking changes are annotated with ☢️, and alpha/beta features with 🐥.
 
 ## Upcoming
 
+This release features significant improvements to [`sq diff`](https://sq.io/docs/diff).
+
+## Added
+
+- Previously `sq diff --data` diffed every row, which could get crazy
+  with a large table. Now the command stops after N differences, where N is controlled by
+  the `--stop` flag, or the new config option [`diff.stop`](https://sq.io/docs/config#diffstop).
+  The default stop-after value is `3`; set to `0` to show all differences.
+
+  ```shell
+  # Stop on first difference
+  $ sq diff @prod.actor @staging.actor --data --stop 1
+  
+  # Stop after 5 differences, using the -n shorthand flag
+  $ sq diff @prod.actor @staging.actor --data -n 5
+  ```
+- [#353]: The performance of `sq diff` has been significantly improved. There's still more to do.
+
 ## Changed
+
+- ☢️ Previously, `sq diff` only exited non-zero on an error. Now, `sq diff` exits `0` when no differences,
+  exits `1` if differences are found, and exits `2` on any error.
+  This aligns with the behavior of [GNU diff](https://www.gnu.org/software/diffutils/manual/):
+
+  ```text
+  Exit status is 0 if inputs are the same, 1 if different, 2 if trouble.
+  ```
 
 - Minor fiddling with the color scheme for some command output.
 
@@ -1131,6 +1157,7 @@ make working with lots of sources much easier.
 [#335]: https://github.com/neilotoole/sq/issues/335
 [#338]: https://github.com/neilotoole/sq/issues/338
 [#340]: https://github.com/neilotoole/sq/pull/340
+[#353]: https://github.com/neilotoole/sq/pull/353
 
 
 [v0.15.2]: https://github.com/neilotoole/sq/releases/tag/v0.15.2